gmail-agent-mcp/requirements.txt
# Gmail API
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.34.0

# Environment management
python-dotenv==1.0.1

# Logging and utilities
pydantic==2.9.2
pydantic-settings==2.6.0

# Development
pytest==8.3.3
black==24.10.0
flake8==7.1.1